Fix: Resolve TypeScript errors for Fight type consistency
This commit addresses TypeScript errors related to the 'Fight' type that were causing build failures.

Changes include:
1. Added missing 'Event' and 'Prediction' interface definitions to 'src/types/mma.ts' and ensured they are exported.
2. Modified 'src/contexts/ufc-context.tsx' to:
   - Import 'Fight', 'Event', and 'Prediction' types from 'src/types/mma.ts', removing local definitions.
   - Update data mapping in 'fetchEvents' and 'fetchFights' to correctly populate 'eventName', 'redCornerName', and 'blueCornerName' string properties on 'Fight' objects, as required by the global 'Fight' type.
   - Used a placeholder for 'eventName' in 'fetchFights' as the direct API response for fights might not include it.
   - Removed incorrect assignments of full Fighter objects to 'redCorner' and 'blueCorner' properties on 'Fight' objects.
   - Refined the mapping of the 'result' property to prevent type errors.

These changes ensure that the data provided through the UFCContext aligns with the centralized type definitions, resolving the type error for 'nextFight.eventName' in 'src/app/(main)/fighters/[id]/page.tsx'.
